+++ CHANGELOG.md
@@ -7,6 +7,13 @@
 - Styling changes:
   - Tags are restyled and positioned below
   - Produce compact CSS in HTML head
+- CLI revamp
+  - Zettelkasten directory is now provided via the `-d` argument.
+    - Its default, `~/zettelkasten`, is used when not specified.
+    - This directory must exist, otherwise neuron will error out.
+  - The output directory is now moved to `.neuron/output` under the Zettelkasten directory.
+  - Added `neuron open` to open the locally generated Zettelkasten site.
+  - `neuron ... rib serve` is now `neuron rib -wS`.
 
 ## 0.2.0.0